reconfigure webpack
